task/requirements.txt
# solver.py 依赖
docker>=7.0.0
python-dotenv>=1.0.0
